backend/ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai>=1.6.1,<2.0.0
Pillow==10.1.0
pydantic==2.5.0
requests==2.31.0
numpy==1.24.3